adding lib files and build complete
